SSX Server CLI and Docker Compose (#8)
* Yarn and TypeScript as project dependencies

Instead of relying on global installation of yarn and typescript, include them as dependencies so they are installed locally to the project and can be pinned to specific versions

* Bump typedoc version

Resolves a peer dependency issues on clean install where two different versions of typedoc are installed alongside each other

* Expose event interface from SSX server

Inherit from EventEmitter for catching and signalling when a login occurs

* Create a SSX server executable

A simple bin script that sets up a basic SSX server and passes configuration through from config file hierachy, cli options, and env vars so that SSX server can be run as a command line utility. Also adds new deps: rc and cors.

* Dockerize SSX server and example dApp

Dockerfiles for running SSX server and the example dApp along with docker compose configuration for starting them both

* Include Docker use instructtion in README

Updates the README to use the `npm` version of the "useful commands" (linked now to use the internal yarn dependency). Adds a docker section describing the environment file and steps for running the e2e example in docker compose.

* Delete package-lock.json

* Use local version of yarn

Updates the GitHub CI workflow to use the latest Node LTS release and the NPM version of ci script

* Remove redundant build/clean ci step

This is already happening as a postinstall hook

* Fix README typos

Replaces instances of "dApp" with the preexisting nomenclature of "dapp"

* Reduce size of Docker containers

Don't install unnecessary dependencies and don't move over node_modules since we are going to install them in the image anyway. Change the base image from debian to node since we don't need to install rust, cargo, python, etc as originally thought.

* Adding back yarn.locl

Resolving upstream merge conflicts ended with removing the yarn.lock so adding it back

Co-authored-by: Gregory Rocco <gregoryrocco1226@gmail.com>